:construction: #6 redesigining Entity and Usecases
