+++ src/Usecase/StockOperator.hs
@@ -1,5 +1,19 @@
 {-# LANGUAGE DeriveGeneric #-}
 
+{-
+このファイルは多分もう使わない。
+Condition --> Usecase.StockCondition
+StockStorable --> Entity.Intenal.StockStore （名前は変わっている）
+で、あと、ここにあったStockStorableはインタラクターと役割がかぶっていたので、
+Usecase.StockStoreAdapterを作って、ここに、StockStoreの実装？でもそうするなら、
+インテーフェイスもユースケース側にあった方が良いので、StockStoreはやはりユースケースなのか。
+この辺りもう一度考える。
+- エンティティとして提供するメソッドのインターフェイスと実装
+- ユースケースとして提供するインターフェイスと実装とインタラクター
+- コントローラーとして提供するドメインインテーフェイスとプラグインインターフェイスのアダプター
+
+
+-}
 module Usecase.StockOperator
   ( Condition(..)
   , StockStorable(..)
@@ -8,23 +22,11 @@ module Usecase.StockOperator
   , StockId
   ) where
 
-import           Data.Aeson   ( FromJSON, ToJSON )
-import           Data.Time    ( Day )
-import           Entity.Stock ( PriceType (..), Stock (..), StockId )
+import           Data.Aeson             ( FromJSON, ToJSON )
+import           Data.Time              ( Day )
+import           Entity.Stock           ( PriceType (..), Stock (..), StockId )
 import           GHC.Generics
-
--- for stock server
-data Condition =
-  Condition
-    { condPriceType :: PriceType
-    , condStartTime :: Day
-    , condEndTime   :: Day
-    }
-  deriving (Eq, Generic, Show)
-
-instance FromJSON Condition
-
-instance ToJSON Condition
+import           Usecase.StockCondition ( Condition (..) )
 
 class StockStorable storage where
   get :: storage -> StockId -> Stock -- get all data of the stock ID
